Made Changes Requested by @Behemyth code review
Moved tests to separate directory.  Fixed importing templated functions.
NOTE: can't build/run project (don't have Nvidia GPU) so @Behemyth
please double check it works.  Especially the import path of Settinhs.h
in  the test file.
